Change "this <repeater>" behaviour for :future and :past contexts
Until now :future and :past contexts would arbitrarily increment or decrement
the start and end respectively of the returned time span to the closest
minute/hour/day. Instead we will return the current value of *NOW* and get rid
of the arbitrariness.

For example, if

  *NOW* => @2006-08-16T14:00:00.000000+05:30

Before this commit, the behaviour was

  (parse "this year" :context :future :guess nil)
  =>
  #<SPAN 2006-08-17T00:00:00.000000+05:30...2007-01-01T00:00:00.000000+05:30>

  (parse "this year" :context :past :guess nil)
  =>
  #<SPAN 2006-01-01T00:00:00.000000+05:30...2006-08-16T00:00:00.000000+05:30>

And after this commit,

  (parse "this year" :context :future :guess nil)
  =>
  #<SPAN 2006-08-16T14:00:00.000000+05:30...2007-01-01T00:00:00.000000+05:30>

  (parse "this year" :context :past :guess nil)
  =>
  #<SPAN 2006-01-01T00:00:00.000000+05:30...2006-08-16T14:00:00.000000+05:30>